requirements.txt
annotated-types==0.7.0
anyio==4.4.0
Authlib==1.3.1
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
distro==1.9.0
flake8==7.1.0
Flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mccabe==0.7.0
openai==1.35.14
packaging==24.1
pluggy==1.5.0
pycodestyle==2.12.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyflakes==3.2.0
pytest==8.2.2
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
